Assigns secret key to user
