+++ app/controllers/application_controller.rb
@@ -16,7 +16,9 @@ class ApplicationController < ActionController::Base
         :email,
         :password,
         :password_confirmation,
-        :remember_me
+        :remember_me,
+        :secret_key_p,
+        :secret_key_q
       )
     end
