+++ db/migrate/20150121185804_add_secret_key_column_to_user.rb
@@ -1,0 +1,8 @@
+class AddSecretKeyColumnToUser < ActiveRecord::Migration
+  def change
+    add_column :users, :secret_key_p, :integer
+    add_column :users, :secret_key_q, :integer
+    add_column :users, :public_key_m, :integer
+    add_column :users, :public_key_k, :integer
+  end
+end
