+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150116182710) do
+ActiveRecord::Schema.define(version: 20150121185804) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -47,6 +47,10 @@ ActiveRecord::Schema.define(version: 20150116182710) do
     t.string   "username",                            null: false
     t.string   "first_name",                          null: false
     t.string   "last_name",                           null: false
+    t.integer  "secret_key_p"
+    t.integer  "secret_key_q"
+    t.integer  "public_key_m"
+    t.integer  "public_key_k"
   end
 
   add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
